Reduce speed and add default sprite for characters

+++ src/mapping/Map.java
@@ -176,18 +176,18 @@ public class Map {
 	}
 
 	public void display(GraphicsContext gc, @NotNull Coordinate characterPosition) {
-		int initColumn = (int) (characterPosition.getX()) - 90;
-		int initLine = (int) (characterPosition.getY()) - 60;
+		double initColumn = characterPosition.getX() - 90;
+		double initLine = characterPosition.getY() - 60;
 		double lineOffset = (characterPosition.getY()) - HEIGHT/(2* SIDE);
 		double columnOffset =  (characterPosition.getX()) - WIDTH/(2*SIDE);
-		for (int column = initColumn; column < initColumn + 180; column++) {
-			for (int line = initLine; line < initLine + 120; line++) {
+		for (double column = initColumn; column < initColumn + 180; column++) {
+			for (double line = initLine; line < initLine + 120; line++) {
 				if (line < 0 || line >= lines || column < 0 || column >= columns) {
 					continue;
 				}
-				gc.drawImage(spriteSelector(map[column][line]),
-						(column - columnOffset) * SIDE - characterPosition.getX() ,
-						(line - lineOffset) * SIDE - characterPosition.getY() , SIDE, SIDE);
+				gc.drawImage(spriteSelector(map[(int)column][(int)line]),
+						(column - columnOffset) * SIDE - characterPosition.getX(),
+						(line - lineOffset) * SIDE - characterPosition.getY(), SIDE, SIDE);
 			}
 		}
 	}
